+++ packages/orq-rc/RELEASES.md
@@ -2209,3 +2209,13 @@ Based on:
 - [typescript v3.8.0-rc.27] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.27 - packages/orq-rc
+
+## 2025-06-15 12:40:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.563.1 (2.629.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.28] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.28 - packages/orq-rc